+++ frameworks/core/components_ng/base/view_abstract_model.h
@@ -161,7 +161,7 @@ public:
     virtual void SetLayoutPriority(int32_t priority) = 0;
     virtual void SetLayoutWeight(float value) = 0;
     virtual void SetLayoutWeight(const NG::LayoutWeightPair& value) = 0;
-    virtual void SetPixelRound(uint8_t value) = 0;
+    virtual void SetPixelRound(uint16_t value) = 0;
     virtual void SetLayoutDirection(TextDirection value) = 0;
     virtual void SetAspectRatio(float ratio) = 0;
     virtual void ResetAspectRatio() = 0;
