+++ frameworks/core/components_ng/base/view_abstract_model_ng.h
@@ -537,7 +537,7 @@ public:
         ViewAbstract::SetLayoutWeight(value);
     }
 
-    void SetPixelRound(uint8_t value) override
+    void SetPixelRound(uint16_t value) override
     {
         ViewAbstract::SetPixelRound(value);
     }
