+++ frameworks/core/components_ng/layout/layout_property.h
@@ -158,12 +158,12 @@ public:
 
     void UpdateChainWeight(const LayoutWeightPair& value);
 
-    void UpdatePixelRound(uint8_t value)
+    void UpdatePixelRound(uint16_t value)
     {
         pixelRoundFlag_ = value;
     }
 
-    uint8_t GetPixelRound() const {
+    uint16_t GetPixelRound() const {
         return pixelRoundFlag_;
     }
 
@@ -438,7 +438,7 @@ private:
 
     bool usingPosition_ = true;
 
-    uint8_t pixelRoundFlag_  = 0;
+    uint16_t pixelRoundFlag_  = 0;
 
     bool isOverlayNode_ = false;
     Dimension overlayOffsetX_;
